+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * Open natively a special file by jemalloc to avoid deadlock
 * Check native events when XCheck*Event() returns nothing
 * Free ScreenCapture when glx context is destroyed
+* Prevent recursive calls to dlsym (#369)
 
 ## [1.4.0] - 2020-06-19
 ### Added